requirements_test.txt
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
aiobotocore>=2.6.0,<3.0.0
homeassistant>=2025.2.0
